requirements.txt
asgiref==3.8.1
Django==5.1.5
django-countries==7.6.1
django-country-kit==0.0.7
django-formtools==2.5.1
django-phonenumber-field==8.1.0
django-widget-tweaks==1.5.0
phonenumbers==9.0.9
pillow==11.1.0
PyMySQL==1.1.1
sqlparse==0.5.3
typing_extensions==4.12.2
